requirements.txt
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==2.2.2
numpy==1.14.2
Pillow==5.1.0
pkg-resources==0.0.0
protobuf==3.5.2.post1
pyparsing==2.2.0
python-dateutil==2.7.2
pytz==2018.3
PyYAML==3.12
scipy==1.0.1
six==1.11.0
tensorboard-logger==0.1.0
torch==0.3.1
torchvision==0.2.0
